requirements.txt
blinker==1.9.0
click==8.1.8
easyocr==1.7.2
filelock==3.18.0
Flask==3.0.3
fsspec==2025.3.2
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pyclipper==1.3.0.post6
pypdf==5.4.0
pyspellchecker==0.8.2
pytest==8.3.3
pytest-mock==3.14.0
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
setuptools==80.3.0
shapely==2.1.0
sympy==1.14.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
Werkzeug==3.1.3
